+++ apps/docs/tailwind.config.cjs
@@ -5,6 +5,10 @@ module.exports = {
   content: ['./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}'],
   theme: {
     extend: {
+      fontSize: {
+        '2xs': '.625rem',
+        '3xs': '.5rem'
+      },
       fontFamily: {
         sans: ['Inter', 'sans-serif']
       },