+++ packages/extras/src/lib/components/controls/OrbitControls/OrbitControls.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { forwardEventHandlers, T, useFrame, useParent, useThrelte } from '@threlte/core'
-  import { Camera } from 'three'
+  import type { Camera } from 'three'
   import { OrbitControls as ThreeOrbitControls } from 'three/examples/jsm/controls/OrbitControls'
   import { useControlsContext } from '../useControlsContext'
   import type {