+++ packages/extras/src/lib/index.ts
@@ -24,6 +24,7 @@ export { default as InstancedMeshes } from './components/Instancing/InstancedMes
 export { default as SoftShadows } from './components/SoftShadows/SoftShadows.svelte'
 export { default as MeshLineGeometry } from './components/MeshLine/MeshLineGeometry.svelte'
 export { default as MeshLineMaterial } from './components/MeshLine/MeshLineMaterial.svelte'
+export { default as Sky } from './components/Sky/Sky.svelte'
 export { Align }
 
 /**