+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @threlte/core
 
+## 6.0.8
+
+### Patch Changes
+
+- b08a992d: Add error messages to context-based hooks
+
+## 6.0.7
+
+### Patch Changes
+
+- e05c73a: Add @threlte/xr package and bump three to 155
+
 ## 6.0.6
 
 ### Patch Changes